최종 머지 승인 부탁드립니다 !!!!!!!!!!!!!!!!!!!!!!!!!!!

See merge request sw_track/class_03/web_project/team24/hugging!78

+++ src/views/mypage/signOut.js
@@ -2,47 +2,55 @@ import * as Api from "/api.js";
 const list_mom = document.getElementById("list_mom");
 const signOut = document.getElementById("signOut");
 
+signOut.addEventListener("click", createForm);
+
 // 회원탈퇴 폼을 제공해주는 함수
-const createForm = async () => {
+async function createForm() {
   while (list_mom.hasChildNodes()) {
     list_mom.removeChild(list_mom.firstChild);
   }
   list_mom.className = "holder_center";
   const clicked_title = document.getElementById("clicked_title");
   const clicked_descript = document.getElementById("clicked_descript");
-  clicked_title.textContent = "회원탈퇴";
-  clicked_descript.textContent = "회원탈퇴 제출 폼";
-  const form = document.createElement("form");
+  clicked_title.textContent = "";
+  clicked_descript.textContent = "";
+
+  const form = document.createElement("div");
   form.id = "signOut_form";
   form.className = "signOut_form";
   form.innerHTML = `
-    <small>'탈퇴' 입력후 제출</small>
+    <p class="modal_title">탈퇴하기</p>
+    <p class="modal_description">'탈퇴' 입력후 제출</p>
     <input
         type="text"
         id="signout_accept"
+        class="input"
         placeholder="'탈퇴'입력"
     />
  
     <br>
     <br>
     <div>
-      <a id="canCel">취소</a>
-      <input type="submit" value="제출" />
+    <button id="signout_btn" class="button">탈퇴하기</button>
+    <button id="canCel" class="button">취소</button>
     </div>
     `;
 
   list_mom.appendChild(form);
-  const signOut_form = document.getElementById("signOut_form");
-  signOut_form.addEventListener("submit", submitFrom);
+
+  // 탈퇴 버튼
+  const signout_btn = document.getElementById("signout_btn");
+  signout_btn.addEventListener("click", submitFrom);
+
+  // 취소버튼
   if (document.getElementById("canCel")) {
     document.getElementById("canCel").addEventListener("click", () => {
       canCel();
     });
   }
   console.log("회원탈퇴 폼 생성");
-};
+}
 
-signOut.addEventListener("click", createForm);
 function canCel() {
   setTimeout(() => {
     while (list_mom.hasChildNodes()) {
@@ -50,24 +58,28 @@ function canCel() {
     }
   }, 1);
   list_mom.className = "";
-  const clicked_title = document.getElementById("clicked_title");
-  const clicked_descript = document.getElementById("clicked_descript");
-  clicked_title.textContent = "";
-  clicked_descript.textContent = "";
 }
 
 const submitFrom = async (e) => {
   e.preventDefault();
-  if (
-    confirm("정말 탈퇴하시겠습니까? \n 탈퇴시 모든 정보는 복구가 불가능합니다.")
-  ) {
-    const body = { accept: document.getElementById("signout_accept").value };
-    const res = await Api.delete("/api/users", "", body);
-    console.log(res);
-    sessionStorage.clear();
-    alert("그동안 감사했습니다.");
-    window.location.href = "/";
-    return;
+
+  const acceptMsg = document.getElementById("signout_accept").value;
+
+  if (acceptMsg === "탈퇴") {
+    if (
+      confirm("정말 탈퇴하시겠습니까?\n탈퇴시 모든 정보는 복구가 불가능합니다.")
+    ) {
+      const body = { accept: acceptMsg };
+      const res = await Api.delete("/api/users", "", body);
+      console.log(res);
+      localStorage.removeItem("loggedIn");
+      localStorage.removeItem("token");
+      localStorage.removeItem("refreshToken");
+      alert("탈퇴되었습니다.\n그동안 감사했습니다.");
+      window.location.href = "/";
+      return;
+    }
+  } else {
+    alert("탈퇴를 원하시면 '탈퇴'를 입력해주세요");
   }
-  alert("휴 다행이에요~ ");
 };
